Add operator tag

+++ src/TagExpression.js
@@ -3,17 +3,31 @@ import PropTypes from 'prop-types'
 import { Tag } from './Tag'
 import { NewTag } from './NewTag'
 import styles from './TagExpression.module.css'
+import { OpTag } from './OpTag'
+import { NewOpTag } from './NewOpTag'
 
 const initialState = {
   accepting: 'cond',
-  tags: { allIds: [], byId: {} },
-  newTag: { field: '', operator: '', value: '' }
+  cnt: 11,
+  tags: {
+    allIds: [1, 2, 3, 4, 5], byId: {
+      1: { id: 1, type: 'cond', field: '1', operator: '2', value: '3' },
+      2: { id: 2, type: 'op', value: 'AND' },
+      3: { id: 3, type: 'cond', field: '1', operator: '2', value: '3' },
+      4: { id: 4, type: 'op', value: 'AND' },
+      5: { id: 5, type: 'cond', field: '1', operator: '2', value: '3' },
+    }
+  },
+  newTag: { field: '', operator: '', value: '' },
+  newOp: { value: '' }
 }
+
+const ops = ['AND', 'OR']
 
 function reducer(state, action) {
   const value = action?.value
   const id = action?.id
-  const { tags: { allIds, byId }, newTag, accepting } = state;
+  const { tags: { allIds, byId }, newTag, newOp, cnt } = state;
 
   switch (action.type) {
     case 'update-new-field':
@@ -25,9 +39,10 @@ function reducer(state, action) {
     case 'add-new-tag':
       return {
         ...state,
-        accepting: accepting === 'cond' ? 'op' : 'cond',
-        tags: { allIds: [...allIds, id], byId: { ...byId, [id]: { id, ...newTag } } },
-        newTag: { field: '', operator: '', value: '' }
+        accepting: 'op',
+        tags: { allIds: [...allIds, cnt], byId: { ...byId, [cnt]: { type: 'cond', id: cnt, ...newTag } } },
+        newTag: { field: '', operator: '', value: '' },
+        cnt: cnt + 1
       }
     case 'update-tag-field':
       return {
@@ -53,6 +68,27 @@ function reducer(state, action) {
           byId: { ...state.tags.byId, [id]: { ...byId[id], value: value } }
         }
       }
+    case 'add-new-op':
+      return {
+        ...state,
+        accepting: 'tag',
+        tags: { allIds: [...allIds, cnt], byId: { ...byId, [cnt]: { type: 'op', id: cnt, ...newOp } } },
+        newOp: { value: '' },
+        cnt: cnt + 1
+      }
+    case 'update-new-op':
+      return {
+        ...state,
+        newOp: { value: value },
+      }
+    case 'update-op':
+      return {
+        ...state,
+        tags: {
+          ...state.tags,
+          byId: { ...state.tags.byId, [id]: { ...byId[id], value: value } }
+        }
+      }
     case 'delete-tag': {
       //eslint-disable-next-line
       const { [id]: _, ...otherTags } = byId
@@ -71,7 +107,7 @@ function reducer(state, action) {
 
 export function TagExpression({ fields, values }) {
   const [state, dispatch] = useReducer(reducer, initialState)
-  const { tags: { allIds, byId }, newTag } = state;
+  const { tags: { allIds, byId }, newTag, newOp, accepting } = state;
   console.log(state)
 
   return (
@@ -79,30 +115,50 @@ export function TagExpression({ fields, values }) {
       {
         allIds.map(tagId => {
           const tag = byId[tagId]
-          return (
-            <Tag
-              key={tag.id}
-              fields={fields}
-              values={values}
-              field={tag.field}
-              operator={tag.operator}
-              value={tag.value}
-              handleFieldChange={(val) => dispatch({ type: 'update-tag-field', id: tagId, value: val })}
-              handleOperatorChange={(val) => dispatch({ type: 'update-tag-operator', id: tagId, value: val })}
-              handleValueChange={(val) => dispatch({ type: 'update-tag-value', id: tagId, value: val })}
-              handleDelete={() => dispatch({ type: 'delete-tag', id: tagId })}
-            />
-          )
+          if (tag.type === 'cond') {
+            return (
+              <Tag
+                key={tag.id}
+                fields={fields}
+                values={values}
+                field={tag.field}
+                operator={tag.operator}
+                value={tag.value}
+                handleFieldChange={(val) => dispatch({ type: 'update-tag-field', id: tagId, value: val })}
+                handleOperatorChange={(val) => dispatch({ type: 'update-tag-operator', id: tagId, value: val })}
+                handleValueChange={(val) => dispatch({ type: 'update-tag-value', id: tagId, value: val })}
+                handleDelete={() => dispatch({ type: 'delete-tag', id: tagId })}
+              />
+            )
+          } else {
+            return (
+              <OpTag
+                key={tag.id}
+                options={ops}
+                value={tag.value}
+                handleChange={(val) => dispatch({ type: 'update-op', id: tagId, value: val })}
+              />
+            )
+          }
+
         })
       }
-      <NewTag
-        fields={fields}
-        values={values}
-        dispatch={dispatch}
-        field={newTag.field}
-        operator={newTag.operator}
-        value={newTag.value}
-      />
+      {
+        accepting === 'tag' ?
+          <NewTag
+            fields={fields}
+            values={values}
+            dispatch={dispatch}
+            field={newTag.field}
+            operator={newTag.operator}
+            value={newTag.value}
+          /> :
+          <NewOpTag
+            options={ops}
+            value={newOp.value}
+            dispatch={dispatch}
+          />
+      }
     </div>
   )
 }
